refactor: extract test_helpers tests to sibling file
